MAYA-129538 small fixes
- Fix copyright years,
- Fix header inclusion guards to not use pragma once.
- Move implementation of a function to its own file.
- Put the common text filtering test helper in the usdUtils file.
- Modify tests to use the common function.
