+++ lib/usdUfe/python/wrapCommands.cpp
@@ -1,5 +1,5 @@
 //
-// Copyright 2021 Autodesk
+// Copyright 2023 Autodesk
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
