+++ test/lib/ufe/testEditRouting.py
@@ -11,16 +11,7 @@ import os
 import unittest
 import usdUtils
 from pxr import UsdGeom
-
-#####################################################################
-#
-# Helper to compare strings.
-
-def filterUsdStr(usdSceneStr):
-    '''Remove empty lines and lines starting with pound character.'''
-    nonBlankLines = filter(None, [l.strip() for l in usdSceneStr.splitlines()])
-    finalLines = [l for l in nonBlankLines if not l.startswith('#')]
-    return '\n'.join(finalLines)
+from usdUtils import filterUsdStr
 
 
 #####################################################################
