+++ test/lib/ufe/testReferenceCommands.py
@@ -10,17 +10,7 @@ import ufe
 import unittest
 import usdUtils
 import testUtils
-
-#####################################################################
-#
-# Helper to compare strings.
-
-def filterUsdStr(usdSceneStr):
-    '''Remove empty lines and lines starting with pound character.'''
-    nonBlankLines = filter(None, [l.strip() for l in usdSceneStr.splitlines()])
-    finalLines = [l for l in nonBlankLines if not l.startswith('#')]
-    return '\n'.join(finalLines)
-
+from usdUtils import filterUsdStr
 
 #####################################################################
 #
